backend/requirements.txt
Django==5.2.11
psycopg==3.3.3
psycopg-binary==3.3.3
djangorestframework==3.15.2
flake8==7.1.0
black==24.4.2
isort==5.13.2
python-decouple==3.8
django-cors-headers==4.3.1
pre-commit==3.7.1
